Move env parsing back to tests/compile/test_backend.py .
